% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rp.norm}
\alias{rp.norm}
\title{Painel para Visualização da Distribuição Normal}
\usage{
rp.norm()
}
\value{
Abre uma janela com deslizadores e caixas de seleção para
visualização da influência dos parâmetros, \code{mean} e \code{sd},
na forma da distribuição Normal.
}
\description{
Esta função cria um painel interativo para visualização
da distribuição Normal. Minha primeira função simples para aprender
sobre sistema de versionamento e trabalho colaboritivo git e github.
}
\examples{
\donttest{
##-----------------------------------------------------------------------------

require(rpanel)
rp.norm()

}
}

